requirements.txt
asgiref==3.2.7
Django==3.0.5
pytz==2019.3
sqlparse==0.3.1
django-crispy-forms==1.9.0
Pillow==7.1.1
django-cleanup==4.0.0
django_nose==1.4.6
coverage==5.1
selenium==3.141.0
webdrivermanager==0.8.0
python-dotenv==0.13.0
mysqlclient==1.4.6
djangorestframework==3.11.0
markdown==3.2.2
django-filter==2.2.0